refactor(feature/10.adicionando-recomendacoes-projeto): alterando para BigDecimal

+++ AluGames/src/main/kotlin/br/com/luizlemetech/alugames/modelo/Aluguel.kt
@@ -1,10 +1,12 @@
 package br.com.luizlemetech.alugames.modelo
 
+import java.math.BigDecimal
+
 data class Aluguel(
     val gamer: Gamer,
     val jogo: Jogo,
     val periodo: Periodo) {
-    val valorDoAluguel: Double = gamer.plano.obterValor(this)
+    val valorDoAluguel: BigDecimal = gamer.plano.obterValor(this)
 
     override fun toString(): String {
         return "Aluguel do ${jogo.titulo} por ${gamer.nome} pelo valor R$$valorDoAluguel"
